+++ pfm/src/main/java/com/thesis/pfm/repository/CustomerBudgetRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
 public interface CustomerBudgetRepository extends JpaRepository<CustomerBudget, Long> {
 
     // Metodo per trovare un budget esistente per una combinazione di cliente e categoria
-    CustomerBudget findByCustomerEmailAndCategoryId(String customerEmail, Long categoryId);
+    CustomerBudget findByCustomerEmailAndCategoryId(String customerEmail, int categoryId);
 
 }
